Break up db model into multiple files (#1013)
- moved `nexus/src/db/model.rs` into `nexus/src/db/model/mod.rs`
- migrated all DB structures into separate files
