+++ nexus/src/db/model/bytecount.rs
@@ -1,0 +1,65 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+use super::BlockSize;
+use diesel::backend::{Backend, RawValue};
+use diesel::deserialize::{self, FromSql};
+use diesel::pg::Pg;
+use diesel::serialize::{self, ToSql};
+use diesel::sql_types;
+use omicron_common::api::external;
+use serde::{Deserialize, Serialize};
+use std::convert::TryFrom;
+
+#[derive(
+    Copy,
+    Clone,
+    Debug,
+    AsExpression,
+    FromSqlRow,
+    Serialize,
+    Deserialize,
+    PartialEq,
+)]
+#[diesel(sql_type = sql_types::BigInt)]
+pub struct ByteCount(pub external::ByteCount);
+
+NewtypeFrom! { () pub struct ByteCount(external::ByteCount); }
+NewtypeDeref! { () pub struct ByteCount(external::ByteCount); }
+
+impl ToSql<sql_types::BigInt, Pg> for ByteCount {
+    fn to_sql<'a>(
+        &'a self,
+        out: &mut serialize::Output<'a, '_, Pg>,
+    ) -> serialize::Result {
+        <i64 as ToSql<sql_types::BigInt, Pg>>::to_sql(
+            &i64::from(self.0),
+            &mut out.reborrow(),
+        )
+    }
+}
+
+impl<DB> FromSql<sql_types::BigInt, DB> for ByteCount
+where
+    DB: Backend,
+    i64: FromSql<sql_types::BigInt, DB>,
+{
+    fn from_sql(bytes: RawValue<DB>) -> deserialize::Result<Self> {
+        external::ByteCount::try_from(i64::from_sql(bytes)?)
+            .map(ByteCount)
+            .map_err(|e| e.into())
+    }
+}
+
+impl From<ByteCount> for sled_agent_client::types::ByteCount {
+    fn from(b: ByteCount) -> Self {
+        Self(b.to_bytes())
+    }
+}
+
+impl From<BlockSize> for ByteCount {
+    fn from(bs: BlockSize) -> Self {
+        Self(bs.to_bytes().into())
+    }
+}
