+++ nexus/src/db/model/disk_state.rs
@@ -1,0 +1,42 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+use diesel::sql_types;
+use omicron_common::api::external;
+use uuid::Uuid;
+
+#[derive(Clone, Debug, AsExpression)]
+#[diesel(sql_type = sql_types::Text)]
+pub struct DiskState(external::DiskState);
+
+impl DiskState {
+    pub fn new(state: external::DiskState) -> Self {
+        Self(state)
+    }
+
+    pub fn state(&self) -> &external::DiskState {
+        &self.0
+    }
+
+    pub fn is_attached(&self) -> bool {
+        self.0.is_attached()
+    }
+
+    pub fn attached_instance_id(&self) -> Option<&Uuid> {
+        self.0.attached_instance_id()
+    }
+}
+
+/// Conversion from the external API type.
+impl From<external::DiskState> for DiskState {
+    fn from(state: external::DiskState) -> Self {
+        Self(state)
+    }
+}
+/// Conversion to the external API type.
+impl Into<external::DiskState> for DiskState {
+    fn into(self) -> external::DiskState {
+        self.0
+    }
+}
