+++ nexus/src/db/model/generation.rs
@@ -1,0 +1,68 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+use diesel::backend::{Backend, RawValue};
+use diesel::deserialize::{self, FromSql};
+use diesel::pg::Pg;
+use diesel::serialize::{self, ToSql};
+use diesel::sql_types;
+use omicron_common::api::external;
+use serde::{Deserialize, Serialize};
+use std::convert::TryFrom;
+
+#[derive(
+    Copy,
+    Clone,
+    Debug,
+    Eq,
+    Ord,
+    PartialEq,
+    PartialOrd,
+    AsExpression,
+    FromSqlRow,
+    Serialize,
+    Deserialize,
+)]
+#[diesel(sql_type = sql_types::BigInt)]
+#[repr(transparent)]
+pub struct Generation(pub external::Generation);
+
+NewtypeFrom! { () pub struct Generation(external::Generation); }
+NewtypeDeref! { () pub struct Generation(external::Generation); }
+
+impl Generation {
+    pub fn new() -> Self {
+        Self(external::Generation::new())
+    }
+}
+
+impl ToSql<sql_types::BigInt, Pg> for Generation {
+    fn to_sql<'a>(
+        &'a self,
+        out: &mut serialize::Output<'a, '_, Pg>,
+    ) -> serialize::Result {
+        <i64 as ToSql<sql_types::BigInt, Pg>>::to_sql(
+            &i64::from(&self.0),
+            &mut out.reborrow(),
+        )
+    }
+}
+
+impl<DB> FromSql<sql_types::BigInt, DB> for Generation
+where
+    DB: Backend,
+    i64: FromSql<sql_types::BigInt, DB>,
+{
+    fn from_sql(bytes: RawValue<DB>) -> deserialize::Result<Self> {
+        external::Generation::try_from(i64::from_sql(bytes)?)
+            .map(Generation)
+            .map_err(|e| e.into())
+    }
+}
+
+impl From<Generation> for sled_agent_client::types::Generation {
+    fn from(g: Generation) -> Self {
+        Self(i64::from(&g.0) as u64)
+    }
+}
