+++ nexus/src/db/model/snapshot.rs
@@ -1,0 +1,45 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+use super::ByteCount;
+use crate::db::identity::Resource;
+use crate::db::schema::snapshot;
+use crate::external_api::views;
+use db_macros::Resource;
+use serde::{Deserialize, Serialize};
+use uuid::Uuid;
+
+#[derive(
+    Queryable,
+    Insertable,
+    Selectable,
+    Clone,
+    Debug,
+    Resource,
+    Serialize,
+    Deserialize,
+)]
+#[diesel(table_name = snapshot)]
+pub struct Snapshot {
+    #[diesel(embed)]
+    identity: SnapshotIdentity,
+
+    project_id: Uuid,
+    disk_id: Uuid,
+    volume_id: Uuid,
+
+    #[diesel(column_name = size_bytes)]
+    pub size: ByteCount,
+}
+
+impl From<Snapshot> for views::Snapshot {
+    fn from(snapshot: Snapshot) -> Self {
+        Self {
+            identity: snapshot.identity(),
+            project_id: snapshot.project_id,
+            disk_id: snapshot.disk_id,
+            size: snapshot.size.into(),
+        }
+    }
+}
